Fix a problem where on the first startup of netshell, the default admin user didn't get a default container.
The problem here was that when Users.authUser() created the default admin user on the first boot, Containers.createUser() automatically invoked Containers.createContainer(), which had the assumption that the current thread was running as a valid user.  On the first boot of netshell, there are no valid users, so we need to create the user first (remember this will be the default admin user), switch to that user, and then create the default admin user's container.

Subsequent invocations of Containers.createUser() will only come via a valid login session, for which there will be a defined user.

This might have the odd side-effect of setting the user of the sshd thread, where it normally doesn't have one.  I'm not sure how important this is, since that thread is privileged anyway.  Logins for initial users (each of which have their own threads) appear to have the correct ownerships.

This fixes an issue first documented in:

https://github.com/esnet/enos/issues/16