Start implementing a shutdown system call and shell command to call it.
This was ported from the enos source tree, bmah-osgi branch, revision 3c2ecb0acb272fa06e785350f29ffded94331b7d.

This allows us to cleanly shut down the OSGi framework, but it could clearly be used for other things too.

Not sure if the syscall is implemented in the right class; this may need to move.