Add confirmation prompt to shutdown command.
This required a small overhaul of the way that ShellInputStream.setEofHack() gets used.  We make sure that gets turned on when the Shell gets a command-line prompt and turned off at other times.  This fixed (hopefully once and for all) a bug observed in the OSGi shell where the EOF inserted after a newline character caused great confusion...this problem was also observed in other shell commands such as passwd and adduser.  We're trying to move towards a model where setEofHack() is only turned on while getting a command line from the top-level ENOS shell.