Make NetShell an OSGi bundle that can be deployed via Karaf.
Making a bundle is done via the Maven bundle plugin.

Add a BootActivator class that contains an OSGi activator...it
basically invokes the BootStrap.main entry point.

Update dependencies to jython 2.5.3 and ssh-core 0.12.0.  The ssh-core
update required some minor code changes, but makes it consistent with
what's used by the Karaf runtime.

Add some new dependencies in the POM that were implicit before
(jgraphx, jackson-core-asl, jackson-mapper-asl)

Remove all of the Felix embedded and autoprocessor stuff, because the
idea is that the OSGi framework should already be running by the time
we get control.

Remove requirement for jline.internal in NetShellTerminal.  We lose a
little logging functionality by this change.

Change the way that SShd sets a custom IOServiceFactoryFactory to work
more correctly in the case of multiple classloaders.

Add a features.xml file that can be used to make a NetShell feature in
Karaf.  This doesn't quite work for making a KAR file because some of
the JAR files need to be wrapped at deploy-time, and so far we haven't
been able to make this happen for files in a KAR.

The OSGi shell subcommand needs to be made working again.