Checkpoint.  Break out python shell into its own OSGi bundle.
This requires that we add a method to ShellCommandFactory to
unregister a shell command.  The Python bundle activator handles
(among potentially other things) registering and unregistering shell
commands.

The main outstanding issue at this point is that we need to give the
Python classloader access to the Java classes in other bundles.

I'm pretty sure that we don't the uber-dynamic OSGi model of
loading/unloading bundles correctly.

Also we need to figure out a good way to re-implement the netshell CLI
feature where you can just invoke the name of a Python script as a
command.  This functionality needs to be made condition on whether we
have the python shell loaded.  (There are multiple OSGi mechanisms for
dealing with this, it's sort of a matter of picking which one to use.)

Update jython to 2.5.3.