Finally get class loading right from Jython.
When we instantiate a new PythonShell, we give the Python interpreter access to all of the packages exported by any OSGi bundle at the time of its creation.  This doesn't handle the case of new bundles getting loaded later (or indeed of bundles going away); we don't have a good way to deal with that yet..

Fix up a bunch of logging stuff.

Remove explicit import of net.es.netshell.api into the netshell-python modules; this is taken care of by the new 'n' improved class loader.