Fix issues surrounding net.es.netshell.api.SSHConnection that were preventing netshell-kernel from reloading from the karaf bundle cache.
Basically the problem was a dependency on some bouncycastle.org package (org.bouncycastle.openssl).  I'm not sure why this was causing issues, but the package in question was only needed by net.es.netshell.api.SSHConnection.  For the moment, make the resolution of this module optional in the netshell-kernel metadata and remove the JAR files from the netshell-kernel package.

While here, remove some hard-coded hostname, usernames, and passwords from SSHConnection.

This change breaks SSHConnection but it needed some reworking anyway.  And we drastically shorten debugging cycle for OSGi issues because we no longer need to uninstall and reinstall the netshell bundles around a karaf restart.