Fix various problems when initializing the netshell-python module in its activator.
Revision 2225a6eaeb696e49c04b514f7fba4cf5a563b738 causes a (very short-lived) Python shell to be started in the bundle activator, to force loading of a number of JAR files, which can only be done from within a privileged thread.  But at the time this code gets called, there is no defined netshell-level user for the currently running thread.  Remove a few assumptions that this user is actually defined so that the shell can actually get started.

Another approach might be to explicitly assign the thread to a user, although making sure that this user always exists and has the needed privilege might be difficult ("admin" might be the "obvious" user, but what if it gets renamed or has its permissions changed?).  This situation is somewhat similar to what we encounter when we create the initial admin user and its container on first startup.