Re-enable the ability to run Python scripts as shell commands.
This took a bit o' doing because it makes netshell-kernel (weakly and
optionally) depend on netshell-python.  A bunch of OSGi-fu follows.

We handle this by creating an interface and class for a
PythonShellService, which is basically an interface to the existing
PythonShell class.  We make netshell-python register this service in
its activator.

In the Shell class command finding / parsing, we first need to see if
we can find the PythonShellService in the OSGi service registry.  If
so, we use it to find and (if found) invoke the appropriate Python
script.

There are probably more architecturally pure ways to handle this, for
example, by using some of the OSGi dependency injection frameworks.
But this is simple enough for us, and it seems to work without us
needing to rearchitect a bunch of existing code.