Don't install the gogo shell command, it doesn't work.  Also remove commented-out Python shell install.
If we bring back access to the OSGi shell, it will need to be implemented in a very different way from what we did originally, and I'm not sure yet what is the best way to do that.  At least this stops us from offering a command that stands no chance of working.