Cleanup leftover cruft from the initial attempt at OSGi integration.
That was an attempt to embed an OSGi runtime container in a netshell main program, but we abandoned that in favor of completely organizing netshell and enos using bundles managed by Karaf, the same way that ODL is organized.

This commit removes all remaining support for the embedded OSGi stack and the gogo shell command.  It also removes a number of dependencies on the Felix OSGi framework, which might open the door towards making NetShell more agnostic to running within other OSGi frameworks such as Equinox.