Commit the start of a NetShell module that integrates with ODL.
It doesn't really do anything right now except register itself with the ODL framework as a new plug-in.

Note that in order to make this work, we needed to solve a compatibility problem with netshell and the Equinox OSGi framework.  Essentially the embedded sshd server was throwing a runtime exception because a class it needed wasn't visible to the Apache Mina SSHD bundle.  We fix this in the integration instructions by forcing the SSH core bundle to refresh its bundle wirings; this allows it to see one of the other Apache Mina bundles, whose functionality which we depend on.