Next steps of ODL / netshell integration.
Export a publically-callable function that allows us to get a list of all network switches.  This allows us, for example, to get all of the DPIDs of all NetFlow switches.

Treat PacketHandler class as sort of a singleton (there's an assumption, but not really well enforced, that there will be only one of these objects instantiated).  ENOS can then ask the class for the single PacketHandler object (whose class probably needs to be either renamed or refactored).  Otherwise there isn't an easy way for ENOS to find that object.

While here, fix a typo in the Activator that caused SwitchManager dependency to be registered incorrectly.