Upgrade from Jython 2.5.2 to Jython 2.7beta4.
A seemingly-required code change is to set an option in the Python interpreter to force it to not try to import site.py at startup time.  This prevents an error in which the Python interpreter somehow can't find site.py at startup time and errors out.  (It exists in the Lib directory in the jython-standalone JAR file.)  It's not clear at that point whether this behavior is a bug or not.

Note that the org.python.core.Options class has a bunch of static members, so these are effectively global options that apply to all Python interpreters as they're instantiated.  It might be better if all these options could be specified as a parameter to the interpreter being created, but in our particular use case this behavior is fine.

One issue to be worked out is that bnd throws an exception whenever we load Jython for the first time.  It's not clear exactly what the problem is, but Python seems to be working more or less correctly.