Replace netshell-python/.gitignore with .gitignore in the top-level directory, and add a few more entries.
A top-level .gitignore covers all of the subdirectories / subprojects underneath it.  Hopefully we can prevent the prior problems of committing object code (in target) and IntelliJ project files (.idea and *.iml).

A quick review of this commit would be appreciated to make sure that it doesn't result in too much stuff being ignored.