Correctly wire up the netshell-odl-mdsal bundle so that the activator-equivalent gets called at bundle load time.
o Give the initial configuration file a reasonable prefix of "04" (used for sequencing, although it doesn't appear to matter much what we call it in the source tree).  In this file, fix a bunch of wrong namespace problems and get rid of apparently-unrecognized <thread-pool-size></thread-pool-size>, replacing it with a commented-out example from a module snarfed from ODL Helium.  It might be necessary to set a numerically higher prefix to push this module initialization later in the initialization sequence.

o Add attach-artifacts goal in netshell-odl-mdsal POM file to get the initial configuration file into the bundle JAR.  Why the pathname in the <file></file> element doesn't match the directory pathname in the source tree (and why this works) is beyond my comprehension.

o In the features.xml file (which we put in the netshell-kernel project...really should move it to someplace higher-level than that), add a <configfile></configfile> element to grab the initial configuration file from the (some) Maven repo and put it into the Karaf directory tree.  Note that the "etc/opendaylight/karaf" part of the pathname is hard-coded...controller projects can grab this from their parent POM file.  It's not clear how we can easily parameterize this.  This causes the configuration file to be read by the config subsystem (along with a bunch of other metadata), which in turns causes the OdlMdsalModule.createInstance() method to get called.  Note that once the configuration file gets downloaded, there doesn't seem to be a way to remove it or (more usefully) update it.