Commit auto-gen files from yangtools.
I'm a little leery of committing auto-generated files in general, however
we actually need to make changes to some of these (i.e. filling in stub
methods) so there is some rationale for this action.  Not sure what's the
right workflow for regenerating these files if we need to, however.