Fix a problem where we could crash with a null pointer reference.
This showed up in testing where we attempted to iterate over a list whose initializer was commented out.  This showed up in testing when under some circumstances (which are not well understood) the MD-SAL module was getting started, stopped, and started again.

Also move the configuration subsystem stuff later in the initialization sequence and shuffle around some diagnostic output.