Code checkpoint:  Finally succeeded in pushing some flows to the OpenFlow controller.
This works for making a single flow that forwards from one switch port to another.  Obviously that's not the most useful flow, but it at least lets us get some packets flowing in the network.

Next actions are to clean this up and to make it implement the Layer 2 (VLAN / MAC translation) flows we really need.