Checkpoint:  We're able to push a VLAN and/or MAC translating flow to an OVS switch now via MD-SAL.
Adjust Corsa version of this call to match APIs.

Corsa version now uses some code from the openflowplugin module.  I am not sure if this is a good idea or not...we know that plugin will be available but it may not be a great idea to introduce this dependency on it (it's some API function calls to some utility functions that bypass the SAL).