Make flow deletion with the MD-SAL OpenFlow plugin work, although there is some API tweaking left to be done.
For some reason, SalFlowService really really really wants the actual flow being pushed when we go to delete it.  I expected that we should be do this just given the FlowRef, because that contains the path to the flow object in the datastore (so SalFlowService would know what to nuke).  If we don't do this, we end up blowing away all of the flows on the switch.

Try a better way of generating the FlowId for a new flow, to make it unique.  Also creation of the flow and pushing it are now two separate operations because we need both the flow and the flow ID (right now we don't know the flow ID until it gets pushed, although it might be possible to change this).  Note that the Corsa driver only returns the flow ID, and that gets passed as a handle to their deleteFlow() method.  I am not sure how this works for them.

Adjust our deleteFlow() API in our Corsa glue to match what's done for generic OpenFlow (it now takes a Flow object as an argument but ignores it, so null can safely be passed here).