Work around a limitation with OVS where it won't output a packet to its input port.
The solution (mentioned in the release notes to OVS 1.11.0) is, if necessary, to overwrite the in_port field in the metadata with the LOCAL port, so that we can output the packet to anywhere we want.

As a side effect, change OpenFlow methods that used to accept Node IDs to take Nodes instead.  (It's easier to take a Node and determine its Id than to go from an Id to a Node object.)