Get initial flows working on mininet.
This allows us to push the initial "send unmatched stuff to the controller" flow to every switch.  It's not needed in the mininet environment (which only does OpenFlow 1.0, which in turn does this behavior automatically anyway), but it was easier to test it there than in the testbed environment where it really matters.