Make it possible to build netshell without the presence of the Corsa ODL driver.
We do this by splitting up the netshell-odl-corsa module into two modules...an interface module (netshell-odl-corsa-intf) that contains only an interface, and an implementation module (essentially the original netshell-odl-corsa module) that contains all of the function implementation.  The interface module has relatively few dependencies, notably it can be built without the Corsa ODL driver.  The implementation driver does require the Corsa ODL driver.  When building Netshell, the generic controller module only depends on having the interface module, but at runtime will dynamically check for the presence of the implementation class (and its single instance) and obtain a pointer to it as needed.

Still remaining is to structure the build so that we present an option (somehow) whether to build netshell-odl-corsa (the implementation module) or not.  There might be some performance aspects of the way that we dynamically check for the implementation class that need to be addressed or at least investigated.

Lightly build and run tested both with and without Corsa ODL driver functionality.