WIP / Checkpoint commit of SDN Controller, SDN Controller client, and RabbitMQ interface.
This is a first step in trying to slightly decouple ENOS from ODL,
basically putting a RabbitMQ interface between the main part of ENOS
and ODL so that ENOS and ODL can run in two separate processes if we
want to move in that direction.  The design gives enough abstraction
that in theory we could even switch controllers to something like RYU
or ONOS.

SDN Controller runs in the same place as ODL.  SDN Controller runs
with the rest of Netshell/ENOS.  There's a RabbitMQ interface between
the two.  Messages are JSON-encoded structures to set up forwarding,
remove forwarding, add and delete meters, send packets (PACKET_OUT),
and receive packets (PACKET_IN).

Right now, all of these pieces live in the netshell-controller bundle;
these would eventually get broken out into at least two bundles (one to
go with ENOS and the other to go with ODL).