The port name in PACKET_IN notifications crossing the RMQ interface should come from FlowCapableNodeConnector.
Previously, the port name was a stringified version of the NodeConnectorId.  This didn't work because the SdnControllerClient (a.k.a. OpenFlow driver) side of the message needed a port name (i.e. "eth10") but doesn't have ODL knowledge or functionality to obtain it.  So we have to do that conversion before composing the SdnReceivePacketReply message.