Fix a bug where sys.path would grow by one element every time a Python command script was executed from the Netshell CLI, without limit.
Due to dependencies on sys, this needs to run fairly late in the startup, ideally in a way where the path adjustment (adding the user home directory to sys.path) can be done conditionally.  The easiest way to do this is in the system profile.py, so that's where we put it.