Fix a problem where interactive shell sessions would exit or hang at odd times.
The command shell behaves differently depending on whether it is called with arguments (specifying a command to run) or no arguments (implying an interactive shell).  We need to make this determination once only, when the shell is first invoked.  Confusing behavior happened when we re-evaluated this after every iteration of the command processing loop; interactive shells were becoming non-interactive.

Also removed some debugging code which seemed out of place.

Possible fix for the reopened #28.