WIP checkpoint for splitting netshell-controller.
netshell-controller-intf contains common classes for any code that needs to use the ENOS SDN controller interface.

netshell-odl-provider is the portion of the ENOS SDN controller interface that runs in an ODL container.

netshell-odl-controller is the portion of the ENOS SDN controller interfaces that runs in the ENOS container.

The OSGi feature definition file has been moved to its own OSGi bundle, netshell-features, because that needs to be accessible from both the ODL container and the ENOS container.  features.xml reflects the new intended organization of the various bundles.

No changes to package names at this point.

Doesn't compile yet, because the bundle activator for netshell-odl-provider depends on a class loader for OSGi bundles (to find the optional netshell-odl-corsa driver) that as of now only exists in netshell-kernel, but netshell-odl-provider cannot depend on netshell-kernel.  We should probably see if we can use the OSGi service registry instead (and indeed we do this in the case of the optional netshell-python module), although netshell-odl-corsa's use of MD-SAL complicates matters.