Fix one more version string issue.
With this, and a fair amount of hand-tweaking of configuration files (still to be documented) to keep the multiple Karaf instances from interfering with each other, we can fire up parallel ODL and ENOS Karaf instances on the same host and do an SdnPing operation initiated by a Python command in the ENOS Karaf to the ODL provider in the ODL Karaf.

Towards #26.