Figure out a new hack to the problem of generating OpenFlow flow entries that need to forward packets out their input ports.
Our prior hack relied on setting the packetIn field in the packet metadata to some unused / bogus value, so that any valid switch port would be different from the input port during packet processing.  OpenFlow specifically disallows changes to the packetIn field (according to source code comments in openvswitch), and starting in openflowvswitch 2.4.0, this hack no longer works.

The "correct" way to do this is when forwarding the packet out its input port, to forward it to the special IN_PORT.  Doing the hack this way appears to be supported (postulated by code inspection) in all versions of openvswitch that we care about.  It even simplifies the flow entry generation code a bit, so it's an all around win.

This bugfix fixes a problem on the ESnet SDN testbed (issue #33), where openvswitch was updated to 2.5.0, and where we could no longer push broadcast fanout flows to the OVS instances, thus breaking layer 2 broadcast support.  This functionality appears to work correctly now.