Move to a pub/sub model for notifications from ODL to Netshell.
This involves the ODL provider sending notification messages to a RabbitMQ exchange, and having the different OpenFlow controller clients listen on queues attached to that exchange.  This change allows multiple Netshells each running a MP-VPN service instance to be working with a single ODL instance; they all get all of the PACKET_IN (and other, when the exist) notifications.  "Slicer" logic will ensure that each MP-VPN instance ignores those messages that don't involve it.

Note that when upgrading across this boundary, both the ODL and Netshell parts need to be upgraded together (because the RabbitMQ interface is now an exchange, not a single queue).

Lightly tested, no regressions noted for the case of a single MP-VPN service instance.

 In support of esnet/enos#98.