Allow the hostname for the RabbitMQ server to be specified in the Netshell configuration file.
The parameter to be used for this is "messagingHost", and it defaults to "localhost".

The SdnControllerClient class will use this parameter to allow it to contact a RabbitMQ server on another host for sending and receiving messages to the SdnController.  In a two-Karaf setup, this allows the Netshell Karaf to live on a different host from the ODL Karaf.  At the moment we assume the ODL Karaf lives on the same host as its RabbitMQ server.

 In support of esnet/enos#98, in which we might want to have multiple Netshell instances (running in different Karafs on different hosts) talking to the same SdnController / ODL Karaf.